% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuSCI.R
\name{MuSCI}
\alias{MuSCI}
\title{The main algorithm multi-source conformal inference (MuSCI)}
\usage{
MuSCI(
  data.train,
  data.test,
  alpha = 0.1,
  conf.score = "ASR",
  Y.name,
  T.name,
  tgt.site.name,
  ps.Library = "SL.glm",
  m.Library = "SL.glm",
  p1 = 0.5,
  seed = 10012
)
}
\arguments{
\item{data.train}{training data, a data.frame() objective with defined variable names}

\item{data.test}{testing data, which should have the same structure as the training data}

\item{alpha}{desired coverage level of the prediction intervals}

\item{Y.name}{name of outcome in the training and testing data}

\item{T.name}{name of the site variable in training data}

\item{tgt.site.name}{name of the target site}

\item{ps.Library}{SuperLearner library for fitting the propensity score;
use SuperLearner::listWrappers() to see all possible choices}

\item{m.Library}{SuperLearner library for fitting the putative cdf of conformal score;
use SuperLearner::listWrappers() to see all possible choices}

\item{p1}{proportion of training data used for nuisance function training, default value 0.5}

\item{seed}{seed for random sampling when splitting data}
}
\description{
The main algorithm multi-source conformal inference (MuSCI)
}
